feat: add function to add a todo to the database
This commit adds a new function called "add" to the "api" module. This function is responsible for adding a new todo to the database. It uses the Firebase Realtime Database API to push a new record to the "todos" collection under the user's ID. The function takes two parameters: "user" and "deed". It creates a new reference to the "todos" collection using the user's ID, pushes a new record with the "deed" data, and then retrieves the newly added record from the database. The retrieved record is then returned by the function. This function will be used to add new todos to the database when a user adds a new todo in the TodoAdd component.